hooks/requirements.txt
-e git+https://github.com/whitmo/charmhelpers.git#egg=charmhelpers
ansible
path.py
docker-py
